Hoist methods up to `SlotMapOwner`.
